Add file-upload API & file-message Chat API
The chat API now supports filemessages. Therefore a file upload API was
introduced, so that files can be uploaded to the server and the uploaded
files can be downloaded with the help of fileupload respnses, which
contain the download URL. With this functionallity filemessages can be
send in the chat, so that files can be shared among users.